+++ deluge/ui/translations_util.py
@@ -11,31 +11,13 @@ from __future__ import unicode_literals
 
 import gettext
 import locale
-import logging
 import os
 import sys
 
 from six.moves import builtins
 
 import deluge.common
-
-log = logging.getLogger(__name__)
-log.addHandler(
-    logging.NullHandler()
-)  # Silence: No handlers could be found for logger "deluge.util.lang"
-
-
-def set_dummy_trans(warn_msg=None):
-    def _func(*txt):
-        if warn_msg:
-            log.warning(
-                '"%s" has been marked for translation, but translation is unavailable.',
-                txt[0],
-            )
-        return txt[0]
-
-    builtins.__dict__['_'] = _func
-    builtins.__dict__['ngettext'] = builtins.__dict__['_n'] = _func
+from deluge.argparser_base import log, set_dummy_trans
 
 
 def get_translations_path():
